+++ sxm_player/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Christopher Bailey"""
 __email__ = "cbailey@mort.is"
-__version__ = "0.2.0"
+__version__ = "0.2.1"